Increase the number of suggestion results returned
Having a maximum number of 20 suggestions being returned can be
restrictive for some uses, such as browsing possible entries when one
is less aware of the potential result-space.  This change makes the
default number of suggestions returned more extensive.  It also allows
a `limit` parameter to be sent with the GET request to override the
default value of 100 if that amount is too large or too small.